create SharedComponents
